Implement Pay, rework Sandbox and Actor methods
Now Sandbox holds separately the amount of ammo available as cubes and
as powerups and tracks the changes of both AmmoAvailable and Powerups
used.

Powerups still aren't to be used for their effect in a sandbox but they
can be used to pay.

On the Effect generation phase of any action the PayTemplate will always
prompt the user for a list of powerups to use as ammo, however after the
first prompt (unless the payment can't be made by ammo cubes alone) the
payment will default to using ammocubes

In the case of more than the necessary powerups being selected the
payment will take in the order they have been presented