Fix respawn and problem with score keeping
In DominationPointTile the damage added by the spawn point to the actor
is considered a damage inflicted by the Actor on itself

In actor.endTurn the Actor parameter is the Actor whose turn just ended.
Hence the check on the amount of damages taken needs to be done on
"this", not the Actor received as a parameter.

Prior to this change if player B died in player A's turn he'd check
( in B.endTurn(A) ) whether or not A is alive (?!). Then on B's turn all
other players will check whether B is alive, realize he isn't and start
the respawn